+++ Controllers/TopicoController.cs
@@ -48,49 +48,87 @@ namespace MediLab.Controllers
         {
             return View();
         }
+        private bool topicoenuso(string topico,int id=0)
+        {
+            var item = db.Topico.Where(s => s.Nombre.Equals(topico) && (!s.Id.Equals(id))).FirstOrDefault();
+            return (item != null);
+
+        }
+
+
         [HttpPost]
         public ActionResult Create(FormCollection collection)
         { try
             {
                 ResultSet response = new ResultSet();
-                Topico topico = new Topico()
+
+                if (!topicoenuso(collection["Nombre"].Trim()))
                 {
-                    Nombre = collection["Nombre"].Trim(),
-                    Descripcion = collection["Descripcion"].Trim()
+                    Topico topico = new Topico()
+                    {
+                        Nombre = collection["Nombre"].Trim(),
+                        Descripcion = collection["Descripcion"].Trim()
 
 
-                };
-               db.Topico.Add(topico);
-               db.SaveChanges();
-               response.Code = 1;
-               response.Msg = String.Format("Se creó el topico {0}", topico.Nombre);        
-               return RedirectToAction("Index", new RouteValueDictionary(response));
+                    };
+
+                    db.Topico.Add(topico);
+                    db.SaveChanges();
+                    response.Code = 1;
+                    response.Msg = String.Format("Se creó el topico {0}", topico.Nombre);
+                }        
+                else
+                {
+                    response.Code = -1;
+                    response.Msg = String.Format("El tópico {0} ya existe", collection["Nombre"].Trim());
+
+                }
+                return RedirectToAction("Index", new RouteValueDictionary(response));
 
             }
-            catch
+            catch(Exception exc)
             {
                 return View();
             }            
         }
         public ActionResult Edit(int id)
-        {           
-            Topico topico = db.Topico.Where(s => s.Id.Equals(id)).First();            
-            return View(topico);
+        {
+            Topico topico = db.Topico.Where(s => s.Id.Equals(id)).FirstOrDefault();
+            if (topico != null)
+            {
+                return View(topico);
+            }
+            else
+            {
+                ResultSet response = new ResultSet();
+                response.Code = -1;
+                response.Msg = String.Format("El tópico seleccionado ya no existe");
+                return RedirectToAction("Index", new RouteValueDictionary(response));
+            }
         }
         [HttpPost]
         public ActionResult Edit(int id, FormCollection collection)
         {
             try
             {
-                ResultSet response = new ResultSet();                       
-                Topico topico = db.Topico.Where(s => s.Id.Equals(id)).First();
-                topico.Nombre = collection["Nombre"].Trim();
-                topico.Descripcion = collection["Descripcion"].Trim();
-                db.SaveChanges();
-                response.Code = 1;
-                response.Msg = String.Format("Se editó el topico {0}", topico.Nombre);
-                return RedirectToAction("Index", new RouteValueDictionary(response));
-                
+                ResultSet response = new ResultSet();
+                Topico topico = db.Topico.Where(s => s.Id.Equals(id)).FirstOrDefault();
+                if (topico != null)
+                {
+                    topico.Nombre = collection["Nombre"].Trim();
+                    topico.Descripcion = collection["Descripcion"].Trim();
+                    db.SaveChanges();
+                    response.Code = 1;
+                    response.Msg = String.Format("Se editó el topico {0}", topico.Nombre);
+                    return RedirectToAction("Index", new RouteValueDictionary(response));
+                }
+                else
+                {                   
+                    response.Code = -1;
+                    response.Msg = String.Format("El tópico seleccionado ya no existe");
+                    return RedirectToAction("Index", new RouteValueDictionary(response));
+
+                }
             }
             catch
             {
@@ -99,21 +137,40 @@ namespace MediLab.Controllers
         }
         public ActionResult Details(int id)
         {
-            Topico topico = db.Topico.Where(s => s.Id.Equals(id)).First();
-            return View(topico);
+            Topico topico = db.Topico.Where(s => s.Id.Equals(id)).FirstOrDefault();
+            if (topico != null)
+            {
+                return View(topico);
+            }
+            else
+            {
+                ResultSet response = new ResultSet();
+                response.Code = -1;
+                response.Msg = String.Format("El tópico seleccionado ya no existe");
+                return RedirectToAction("Index", new RouteValueDictionary(response));
+
+            }
         }
         public ActionResult Delete(int id)
         {
             try
             {
                 ResultSet response = new ResultSet();
-                Topico topico = db.Topico.Where(s => s.Id.Equals(id)).First();
-                db.Topico.Remove(topico);
-                db.SaveChanges();
-                response.Code = 1;
-                response.Msg = String.Format("Se borró el topico {0}", topico.Nombre);
-                return RedirectToAction("Index", new RouteValueDictionary(response));
-
+                Topico topico = db.Topico.Where(s => s.Id.Equals(id)).FirstOrDefault();
+                if (topico != null)
+                {
+                    db.Topico.Remove(topico);
+                    db.SaveChanges();
+                    response.Code = 1;
+                    response.Msg = String.Format("Se borró el topico {0}", topico.Nombre);
+                    return RedirectToAction("Index", new RouteValueDictionary(response));
+                }
+                else
+                {
+                    response.Code = -1;
+                    response.Msg = String.Format("El tópico seleccionado ya no existe");
+                    return RedirectToAction("Index", new RouteValueDictionary(response));
+                }
             }
             catch
             {
